Reorganize LLEP files into llep/ package
Move llep.py, llep_kernels.py, llep_autotune.py into
torchtitan/distributed/llep/ and docs/llep.md as its README.md.
All external imports (from torchtitan.distributed.llep import ...)
remain unchanged since llep/__init__.py preserves the public API.
